Add files via upload
A classic snake game using Java programming. Creating the game board by making a SCREEN_WIDTH * SCREEN_HEIGHT dimensional array. The snake should begin in the center of the board, and be one block long. Randomly place apples each turn on the game board. Displaying the game board each time the snake moves one block toward the apple. The snake should take a random path each turn to reach the apple. Once the snake eats the apple add one cell length to the tail of the snake. If the snake ever touches itself; end the game. If the snake goes out of bounds; end the game.